requirements.txt
# --- Core Framework ---
streamlit==1.34.0
python-dotenv==1.0.1

# --- HTTP ---
requests==2.31.0

# --- LangChain Ecosystem ---
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-google-genai==0.0.8

# --- Embeddings & Vector DB ---
sentence-transformers==2.6.1
chromadb==0.4.24
huggingface-hub==0.21.4  # Required by HuggingFaceEmbeddings

# --- Google GenAI SDK ---
google-generativeai==0.3.2

# --- Logging & Utilities (usually standard but explicitly safe) ---
tqdm==4.66.2
typing-extensions>=4.5.0
fastapi>=0.110.0